Select JWK by `kid` to get around python-jose bug
Python jose has a bug where it'll crash with a JWK type mismatch when
validating a JWT against a heterogeneous set of JWKs. I opened a PR
against the project to manage that case better in accordance with the
JOSE RFC, but that project is is poorly maintained and so it may be
worthwhile to work around it for now.

Also, it may be worth switching to PyJWT, which supports this use case
just as well and is better maintained.